requirements.txt
dagster>=1.5.0
kafka-python>=2.0.2
fastavro>=1.8.0
confluent-kafka[avro]>=2.1.0
requests>=2.28.0
protobuf>=4.21.0,<6.0
grpcio-tools>=1.50.0
googleapis-common-protos>=1.56.0